% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(
  df,
  x = "col",
  y = "row",
  fill = "val",
  title = NULL,
  show_xlabel = TRUE,
  show_ylabel = TRUE,
  annotations = FALSE,
  scientific_notation = FALSE,
  digits = 1
)
}
\description{
Plot a heatmap
}
\examples{
plot_heatmap(
  df,
  title="Raw Data",
  show_xlabel=TRUE,
  show_ylabel=TRUE,
  annotations=TRUE,
  scientific_notation=FALSE,
  digits=0
)
}
