% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_single_line}
\alias{plot_single_line}
\title{Plot a single line plot with directional error bars}
\usage{
plot_single_line(
  df,
  x = "x",
  y = "y",
  xlabel = NULL,
  ylabel = NULL,
  title = NULL,
  error_y = NULL,
  error_y_minus = NULL,
  show_xlabel = TRUE,
  show_ylabel = TRUE,
  show_legend = TRUE,
  error_bar_width = 1,
  palette = "Set2"
)
}
\description{
Plot a single line plot with directional error bars
}
\examples{
plot_single_line(
  df,
  x='x', y='y',
  xlabel='X Axis', ylabel='Y Axis',
  title='Raw Data',
  error_y='plus_e',  # error_y_minus='minus_e',
  show_legend=FALSE,
  error_bar_width=2,
  palette="Set2"
)
}
