% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(
  df,
  x = "col",
  y = "row",
  fill = "val",
  title = NULL,
  show_xlabel = TRUE,
  show_ylabel = TRUE,
  annotations = FALSE,
  scientific_notation = FALSE
)
}
\description{
plots a heatmap
}
\examples{
plotHeatmap(df, title="Raw Data", show_xlabel=TRUE, show_ylabel=TRUE, scientific_notation=FALSE, annotations=TRUE)
}
