% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLine.R
\name{plotLine}
\alias{plotLine}
\title{plotLine}
\usage{
plotLine(
  df,
  x = "x",
  y = "y",
  error_y = NULL,
  error_y_minus = NULL,
  xlabel = NULL,
  ylabel = NULL,
  title = NULL,
  show_xlabel = TRUE,
  show_ylabel = TRUE,
  show_legend = TRUE,
  palette = "Set1",
  error_bar_width = 1
)
}
\description{
plots a line plot with error bars
}
\examples{
plotLine(df, title="Raw Data")
}
