% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_csv_from_text}
\alias{read_csv_from_text}
\title{Reads and parses csv embedded in .txt files}
\usage{
read_csv_from_text(
  file_path,
  encoding = "UTF-16",
  sep = "\\t",
  skiprows = 0,
  nrows = NULL,
  skipcols = 0,
  ncols = NULL,
  index = NULL,
  columns = NULL,
  numeric = FALSE
)
}
\description{
This is useful for data exported from plate readers
}
\examples{
# for 96-well plates:
df <- read_csv_from_text(
  file_path,
  skiprows=3, nrows=8,
  skipcols=2, ncols=12,
  index=LETTERS[1:8],
  columns=seq(1, 12)
)
}
