% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_love_animals.R
\name{readCsvFromText}
\alias{readCsvFromText}
\title{readCsvFromText}
\usage{
df <- readCsvFromText(
    filePath,
    skiprows=3, nrows=8,
    skipcols=2, ncols=12,
    index=LETTERS[1:8],
    columns=seq(1, 12)
)
}
\arguments{
\item{filePath}{}
     {nrows}{numeric}
     {ncols}{numeric}
     {skiprows}{numeric}
     {skipcols}{numeric}
     {columns}{array}
     {index}{array}
}
\description{
}
