% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{smelt}
\alias{smelt}
\title{Special Melt}
\usage{
smelt(df, rowname = "row", colname = "col", valname = "val")
}
\description{
Convert a dataframe from a table to long format
This is an alternative to melt that doesn't throw errors
See: \href{https://stackoverflow.com/questions/28355377/how-to-add-index-of-a-list-item-after-melt-in-r}{Stack Overflow link}
}
