% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCsvFromText.R
\name{readCsvFromText}
\alias{readCsvFromText}
\title{readCsvFromText}
\usage{
readCsvFromText(
  filePath,
  encoding = "UTF-16",
  sep = "\\t",
  skiprows = 0,
  nrows = NULL,
  skipcols = 0,
  ncols = NULL,
  index = NULL,
  columns = NULL
)
}
\description{
Convenience function for reading text files where csv may be embedded
Use the following settings for 96-well plates:
}
\examples{
df <- read_csv_from_txt(
           filePath,
           skiprows=3, nrows=8,
           skipcols=2, ncols=12,
           index=LETTERS[1:8],
           columns=seq(1, 12)
)
}
