% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smelt.R
\name{smelt}
\alias{smelt}
\title{special melt}
\usage{
smelt(df, rowname = "row", colname = "col", valname = "values")
}
\description{
convert a dataframe from a table to long format \cr
this is an alternative to melt that doesn't throw errors \cr
See: \href{https://stackoverflow.com/questions/28355377/how-to-add-index-of-a-list-item-after-melt-in-r}{Stack Overflow link}
}
